backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
icalendar==5.0.10
python-dateutil==2.8.2
pydantic-settings==2.0.3
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
aiofiles==23.2.1
openai==1.2.4
pillow==10.1.0
SpeechRecognition==3.10.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
schedule==1.2.1
pytz==2023.3.post1 